Fix displayed amounts scaling bug and align payment schemas with BigInteger minor units

+++ app/schemas/payment.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from uuid import UUID
 
 from pydantic import BaseModel, ConfigDict, Field
@@ -6,7 +5,7 @@ from pydantic import BaseModel, ConfigDict, Field
 
 class PaymentCreate(BaseModel):
     order_id: str = Field(min_length=1, max_length=255)
-    amount: Decimal = Field(gt=0)
+    amount: int = Field(gt=0)
     currency: str = Field(
         default="INR",
         min_length=3,
@@ -26,7 +25,7 @@ class PaymentResponse(BaseModel):
     id: UUID
     merchant_id: UUID
     order_id: UUID
-    amount: Decimal
+    amount: int
     currency: str
     payment_method: str
     transaction_id: str = Field(validation_alias="razorpay_payment_id")
